Cleaned up the RandomizerMessageBox screen and added maximum text
